style) print_error 위치 config -> log 변경
