+++ streamlit/pages/02_Extra_Category.py
@@ -6,7 +6,8 @@ from fastapi import status
 
 # custom
 from utils.attribute import get_music_category
-from utils.config import add_logo, delete_another_session_state, set_page, print_error
+from utils.config import add_logo, delete_another_session_state, set_page
+from utils.log import print_error
 from utils.generator import make_category_request_json, make_audio_data
 from streamlit_space import space
 from models.Content import MusicContent
